requirements.txt
pandas>=2.2,<3.0
pyserial>=3.5,<4.0
watchdog>=4.0,<7.0
fastapi>=0.115,<1.0
uvicorn>=0.32,<1.0
httpx>=0.27,<1.0
pythonnet>=3.0.3,<4.0
pytest>=8.0,<9.0
